Api/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blockbuster==1.5.24
build==1.2.2
buildozer @ git+https://github.com/misl6/buildozer.git@ee16a4d25f6380fa3896f5c9bbc284ffcf5d2af4
cachetools==5.5.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==44.0.3
Cython==0.29.33
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
docopt==0.6.2
docutils==0.20.1
durationpy==0.8
executing==2.2.0
fastapi==0.115.0
fastjsonschema==2.21.1
filelock==3.12.2
flatbuffers==24.3.25
forbiddenfruit==0.1.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.4
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
Kivy==2.2.1
Kivy-Garden==0.1.5
kubernetes==31.0.0
langchain==0.3.1
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.66
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.5.0
langgraph-api==0.2.72
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.3
langgraph-prebuilt==0.5.1
langgraph-runtime-inmem==0.3.3
langgraph-sdk==0.1.72
langsmith==0.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pep517==0.13.0
pexpect==4.8.0
platformdirs==3.10.0
posthog==3.6.6
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic-settings==2.5.2
pydantic_core==2.33.2
Pygments==2.16.1
pyjnius==1.5.0
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-for-android==2023.5.21
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.1
rpds-py==0.25.1
rsa==4.9
Send2Trash==1.8.3
sh==1.14.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.34
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.38.6
structlog==25.4.0
sympy==1.13.3
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.20.0
toml==0.10.2
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
truststore==0.10.1
typer==0.12.5
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.31.0
uvloop==0.20.0
virtualenv==20.24.3
voila==0.5.8
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
widgetsnbextension==4.0.14
wrapt==1.16.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
zstandard==0.23.0
